Update Microsoft.Windows.SDK.CPP latest to 10.0.20348.19

+++ Microsoft.Windows.SDK.CPP/latest/Include/um/dxtmpl.h
@@ -121,7 +121,7 @@ struct CDXPlex    // warning variable length structure
 
     static CDXPlex* PASCAL_INLINE Create( CDXPlex*& pHead, UINT nMax, UINT cbElement )
     {
-	if ((nMax * cbElement) > (SIZE_MAX - sizeof(CDXPlex)))
+	if ((SIZE_T)(nMax * cbElement) > (SIZE_MAX - sizeof(CDXPlex)))
 	    return NULL;
         CDXPlex* p = (CDXPlex*) new BYTE[sizeof(CDXPlex) + nMax * cbElement];
         if (p == NULL)
